修复逻辑

+++ web/admin/del.jsp
@@ -8,14 +8,19 @@
 <%@ page contentType="text/html;charset=UTF-8" language="java" %>
 <%@include file="../database.jsp"%>
 <%
-    String targetID=request.getParameter("id");
-    Statement sql;
-    int rs;
-    sql=conn.createStatement();
-    rs=sql.executeUpdate("DELETE FROM `url` WHERE id='"+targetID+"'");
-    if(rs==1){
-        response.sendRedirect("admin.jsp?c=1");
-    }else {
-        response.sendRedirect("admin.jsp?c=0");
+    Cookie[] cookies = null;
+    cookies=request.getCookies();
+    if(cookies.length-1!=0){
+        String targetID=request.getParameter("id");
+        Statement sql;
+        int rs;
+        sql=conn.createStatement();
+        rs=sql.executeUpdate("DELETE FROM `url` WHERE id='"+targetID+"'");
+        if(rs==1){
+            response.sendRedirect("admin.jsp?c=1");
+        }else {
+            response.sendRedirect("admin.jsp?c=0");
+        }
     }
+
 %>
